Found some bugs
Was explaining some C++ elements to a friend and discovered some bugs like the reference counter not properly handling a move to itself, and placement new of the ReferenceCounter Underlying Type potentially invoking a user created new. I've got a lot more `new`'s to fix, but I'll get to that later.