Added Base64 support, improved ConPrinter
Added and tested Base64 conversion. This will be used to implement UTF-7.
Fixed some more templates being prefixed with underscores.
Changed ConPrinter from a class to namespace. This was to remove TPrinter, and the generic `Print` function. This was causing problems with strings that were declared as arrays. Changing this to a namespace allows for users to be able to add custom `Print` routines without the need for TPrinter. Technically this should've worked if the `Print` functions were in the global scope.
Added support for padded integers to ConPrinter.
Added support for scientific form, and precision limiting of floating point numbers to ConPrinter.