Smart Pointer rework
Forced usage of atomic increment and decrement.
Fixed checks on `->ObjPtr()` as those are guaranteed to always exceed, even when the RCDO/SWRC is null.
Cleaned up checks on `_rcdo` and `_swrc` to now be optional (they should never be truly null anymore).
Set all `_rcdo` and `_swrc` to `Null()` instead of `nullptr`.
Added more checks to validate a reference count due to the nullability changes.
Changed the order of strong pointers to destruct then decrement instead of decrement then release which was causing problems with weak pointers freeing memory too early.