Removed LLVM IR, generalized string formatter, added all char types to stringbuilder
Remove the TauUtilsStatic Package and LLVM IR. It wasn't being used, and only caused problems when trying to compile.
Generalized the ConPrinter::Print function to a formatter. This will allow me soon to implement a Format function that outputs a DynString.
Added all character types to StringBuilder. This is going to be for the format function. Still need to handle my string types, but raw string pointers are working, so the rest is trivial.