Refactored some files, added PathSanitizer and Base32
Refactored the PageAllocator to move the PageSize function to the header, as it operated on static class members. Moved the page size initialization code to a platform specific file, but left the static variables in a common file.
Copied and refactored the PathSanitizer from TauEngine's ResourceLib.
Added Base32 support, which is required for TOTP tokens.
Made some of the string transformation functions constexpr.
Added AlignToAny, a set of alignment functions that can work on any alignment value instead of just Base2 values. These are useful when you need something to always be a multiple of say 3, but they come at the cost of requiring integer division (actually modulus, but that is still performed by the idiv instruction).