Various minor fixes for macos
The num types min and max declarations were using suffixes only available in MSVC.
Clang complains about the first prototype of StringCast not being declared delete.
mbsrtowcs_s and wcsrtombs_s do not exist outside of MSVC.
CountTrailingZero and CountLeadingZero was not correct.
Wrapped the immintrin.h header in Dragon4 for when on ARM. It wasn't actually being used anyways.